🍄pacman rotations now posting and getting from all of clients but + ..
difficult to say what were coming from pacman rotation ?  a rotation belong a single pacman, but which pacman below this positions ? an if server says to client hello there a positions message for you. client says "oh okay okay" but which pacman  I write? and what is a user ? what is a pacman 😆

So we need a Id based pacman object creator function.
if new client enters game, a id which belong to user, need to be trigged and give a new object about pacman